+++ src/radiobuttongroup.h
@@ -8,7 +8,7 @@ class Checkbox;
 
 class RadioButtonGroup : public Widget {
   public:
-    RadioButtonGroup(Position pos, size_t width, const std::vector<std::string> &items);
+    RadioButtonGroup(Position pos, int width, const std::vector<std::string> &items);
 
     std::function<void(int index, Checkbox &button)> on_selected = {};
     std::vector<std::shared_ptr<Checkbox>> radio_buttons = {};
